+++ baybe/utils/sampling_algorithms.py
@@ -1,17 +1,21 @@
 """A collection of point sampling algorithms."""
 
 import warnings
+from collections import Counter
+from collections.abc import Collection
 from enum import Enum
 from typing import Literal
 
 import numpy as np
 import pandas as pd
 
+from baybe.utils.basic import is_all_instance
+
 
 def farthest_point_sampling(
     points: np.ndarray,
     n_samples: int = 1,
-    initialization: Literal["farthest", "random"] = "farthest",
+    initialization: Literal["farthest", "random"] | Collection[int] = "farthest",
     random_tie_break: bool = True,
 ) -> list[int]:
     """Select a subset of points using farthest point sampling.
@@ -27,10 +31,13 @@ def farthest_point_sampling(
             dimensionality of the points.
         n_samples: The total number of points to be selected.
         initialization: Determines how the first points are selected:
+
             * ``"farthest"``: The first two selected points are those with the
               largest distance. If only a single point is requested, a deterministic
               choice is made based on the point coordinates.
             * ``"random"``: The first point is selected uniformly at random.
+            * Indices: Points corresponding to these indices will be pre-selected.
+
         random_tie_break: Determines if points are chosen deterministically or randomly
             in equidistant situations. If ``True``, a random point is selected from the
             candidates, otherwise the first point is selected. For non-equidistant
@@ -40,20 +47,61 @@ def farthest_point_sampling(
         A list containing the positional indices of the selected points.
 
     Raises:
+        ValueError: If the number of requested samples is less than 1.
         ValueError: If the provided array is not two-dimensional.
         ValueError: If the array contains no points.
         ValueError: If the input space has no dimensions.
-        ValueError: If an unknown method for initialization is specified.
+        ValueError: Indices for initialization are not unique.
+        ValueError: More initialization indices than available points are provided.
+        ValueError: Initialization indices are out of bounds.
+        ValueError: Unknown initialization method.
+        ValueError: More points are requested than available.
     """
+    # Input validation
+    if n_samples < 1:
+        raise ValueError(
+            f"The number of requested samples must be at least 1. "
+            f"Provided: {n_samples=}."
+        )
+
     if (n_dims := np.ndim(points)) != 2:
         raise ValueError(
             f"The provided array must be two-dimensional but the given input had "
             f"{n_dims} dimensions."
         )
+
     if (n_points := len(points)) == 0:
         raise ValueError("The provided array must contain at least one row.")
+
     if points.shape[-1] == 0:
         raise ValueError("The provided input space must be at least one-dimensional.")
+
+    if isinstance(initialization, Collection) and is_all_instance(initialization, int):
+        if duplicates := {k for k, v in Counter(initialization).items() if v > 1}:
+            raise ValueError(
+                f"The provided collection of initialization indices must be unique but "
+                f"contains duplicates: {duplicates}"
+            )
+        if len(initialization) > n_points:
+            raise ValueError(
+                f"The number of provided initialization indices "
+                f"({len(initialization)}) cannot be larger than the total number of "
+                f"points provided ({n_points})."
+            )
+        if problematic_indices := [
+            idx for idx in initialization if not (0 <= idx < n_points)
+        ]:
+            raise ValueError(
+                f"The provided collection of initialization indices must be within the "
+                f"range of available points (0 to {n_points - 1}) but contains "
+                f"out-of-bounds indices: {problematic_indices}"
+            )
+    elif initialization not in {"farthest", "random"}:
+        raise ValueError(
+            f"Unknown initialization type. Expected 'farthest', 'random', or a "
+            f"collection of integers. Provided: {initialization=}"
+        )
+
     if n_samples > n_points:
         raise ValueError(
             f"The number of requested samples ({n_samples}) cannot be larger than the "
@@ -87,8 +135,10 @@ def farthest_point_sampling(
         )
         if n_samples == 1:
             return [sort_idx[selected_point_indices[0]]]
-    else:
-        raise ValueError(f"unknown initialization recommender: '{initialization}'")
+    elif isinstance(initialization, Collection) and is_all_instance(
+        initialization, int
+    ):
+        selected_point_indices = list(initialization)
 
     # Initialize the list of remaining points
     remaining_point_indices = list(range(n_points))
@@ -102,15 +152,15 @@ def farthest_point_sampling(
 
         # Find for each candidate point the smallest distance to the selected points
         min_dists = np.min(dist, axis=1)
+        max_val = np.max(min_dists)
+        max_indices = np.where(min_dists == max_val)[0]
 
         if random_tie_break:
             # Select a random point that has the "largest smallest distance"
-            max_val = np.max(min_dists)
-            max_indices = np.where(min_dists == max_val)[0]
             choice = np.random.choice(max_indices)
         else:
-            # Choose the first point with the "largest smallest distance"
-            choice = np.argmax(min_dists)
+            # Choose the last point with the "largest smallest distance"
+            choice = max_indices[-1]
         selected_point_index = remaining_point_indices[choice]
 
         # Add the chosen point to the selection